+++ app/src/main/java/com/example/marvelapp/data/service/response/CharacterResponse.kt
@@ -2,7 +2,7 @@ package com.example.marvelapp.data.service.response
 
 import com.google.gson.annotations.SerializedName
 
-open class CharacterResponse(
+ class CharacterResponse(
     @SerializedName("id") val id: Int,
     @SerializedName("name") val name: String,
     @SerializedName("description") val description: String,
