--- app/src/main/java/com/example/marvelapp/domain/service/MarvelServiceCharacterDetail.kt
@@ -1,9 +1,0 @@
-package com.example.marvelapp.domain.service
-
-import com.example.marvelapp.domain.entity.CharacterDetailEntity
-import com.example.marvelapp.domain.entity.CharacterEntity
-import io.reactivex.rxjava3.core.Observable
-
-interface MarvelServiceCharacterDetail {
-    fun getSingleCharacter(id :Int): Observable<CharacterDetailEntity>
-}
