+++ app/src/main/java/com/example/marvelapp/domain/usecase/GetSingleCharacterUseCaseImpl.kt
@@ -1,15 +1,14 @@
 package com.example.marvelapp.domain.usecase
 
 import com.example.marvelapp.domain.entity.CharacterDetailEntity
-import com.example.marvelapp.domain.service.MarvelServiceCharacterDetail
+import com.example.marvelapp.domain.service.MarvelService
 import io.reactivex.rxjava3.core.Observable
 
 interface GetSingleCharacterUseCase {
     fun call(id: Int): Observable<CharacterDetailEntity>
 }
 
-class GetSingleCharacterUseCaseImpl constructor(private val marvelService: MarvelServiceCharacterDetail) :
+class GetSingleCharacterUseCaseImpl constructor(private val marvelService: MarvelService) :
     GetSingleCharacterUseCase {
-    override fun call(id:Int): Observable<CharacterDetailEntity> = marvelService.getSingleCharacter(id)
-
+    override fun call(id: Int): Observable<CharacterDetailEntity> = marvelService.getSingleCharacter(id)
 }
