+++ app/src/main/java/com/example/marvelapp/presentation/mvp/view/MarvelAppView.kt
@@ -23,7 +23,7 @@ class MarvelAppView(activity: Activity, private val binding: ActivityMainBinding
     }
 
     override fun showError() {
-        Toast.makeText(context, R.string.conection_not_established, Toast.LENGTH_SHORT).show()
+        Toast.makeText(context, R.string.connection_not_established, Toast.LENGTH_SHORT).show()
     }
 
     override fun showLoading() {
