+++ app/src/main/res/values/strings.xml
@@ -1,7 +1,7 @@
 <resources>
     <string name="app_name">MarvelApp</string>
-    <string name="conection_established">"The connection with the server has been established"</string>
-    <string name="conection_not_established">"Can not conect with the server"</string>
+    <string name="connection_established">"The connection with the server has been established"</string>
+    <string name="connection_not_established">"Can not connect with the server"</string>
     <string name="character_image">Character Image</string>
     <string name="description">Description</string>
     <string name="appearances_label">Number of appearances in comics</string>
